Updating translations for NL and BG (#345)
This PR adds the strings for "enemy panel", "iceland" and "pangaea". It
also organises the json file in a different way due to the start of use
of Crowdin (it doesn't interfere in any way with the functionality of
the translations). It also edits the existing translations for the
languages.

Credits:
- Bulgarian - Nikola123
- Dutch - cldprv, tryout33